shit_poster/requirements.txt
google-cloud-pubsub>=2.13.0
google-auth>=2.16.0
python-dotenv>=0.19.0
PyGithub>=1.58.2
